v2/requirements.txt
numpy==1.22.0
opencv_python==4.5.5.62
Pillow==9.0.0
scipy==1.7.3
